Rename apiWrapper to fetchWrapper
