+++ client/src/PlaylistGrid.tsx
@@ -3,7 +3,7 @@ import React, { CSSProperties, useEffect } from 'react'
 import { classes, colors } from './styles'
 import { useHover } from './useHover'
 import { Link } from 'react-router-dom'
-import { useResource, apiWrapper, Resource } from './apiWrapper'
+import { useResource, fetchWrapper, Resource } from './fetchWrapper'
 
 
 
@@ -15,7 +15,7 @@ export const usePlaylists = (): Resource<Playlists> => {
   useEffect(() => {
     (async () => {
       // setter({ loading: true })
-      const response = await apiWrapper('/api/playlists/')
+      const response = await fetchWrapper('/api/playlists/')
       setter({
         loading: false,
         ...response,
